Refactor packages of ParameterKind, TypeDescKind and MethodDescriptor
