+++ compiler/ballerina-lang/src/main/java/io/ballerina/compiler/api/impl/types/BallerinaFutureTypeDescriptor.java
@@ -20,7 +20,7 @@ package io.ballerina.compiler.api.impl.types;
 import io.ballerina.compiler.api.ModuleID;
 import io.ballerina.compiler.api.types.BallerinaTypeDescriptor;
 import io.ballerina.compiler.api.types.FutureTypeDescriptor;
-import io.ballerina.compiler.api.types.TypeDescKind;
+import io.ballerina.compiler.api.types.util.TypeDescKind;
 import org.wso2.ballerinalang.compiler.semantics.model.types.BFutureType;
 
 import java.util.Optional;
