+++ compiler/ballerina-lang/src/main/java/io/ballerina/compiler/api/impl/types/BallerinaRecordTypeDescriptor.java
@@ -21,7 +21,7 @@ import io.ballerina.compiler.api.ModuleID;
 import io.ballerina.compiler.api.types.BallerinaTypeDescriptor;
 import io.ballerina.compiler.api.types.FieldDescriptor;
 import io.ballerina.compiler.api.types.RecordTypeDescriptor;
-import io.ballerina.compiler.api.types.TypeDescKind;
+import io.ballerina.compiler.api.types.util.TypeDescKind;
 import org.wso2.ballerinalang.compiler.semantics.model.types.BRecordType;
 
 import java.util.List;
