+++ compiler/ballerina-lang/src/main/java/io/ballerina/compiler/api/impl/types/BallerinaStreamTypeDescriptor.java
@@ -20,7 +20,7 @@ package io.ballerina.compiler.api.impl.types;
 import io.ballerina.compiler.api.ModuleID;
 import io.ballerina.compiler.api.types.BallerinaTypeDescriptor;
 import io.ballerina.compiler.api.types.StreamTypeDescriptor;
-import io.ballerina.compiler.api.types.TypeDescKind;
+import io.ballerina.compiler.api.types.util.TypeDescKind;
 import org.wso2.ballerinalang.compiler.semantics.model.types.BStreamType;
 
 import java.util.List;
