+++ compiler/ballerina-lang/src/main/java/io/ballerina/compiler/api/impl/types/TypeBuilder.java
@@ -21,14 +21,13 @@ package io.ballerina.compiler.api.impl.types;
 import io.ballerina.compiler.api.ModuleID;
 import io.ballerina.compiler.api.impl.BallerinaModuleID;
 import io.ballerina.compiler.api.impl.LangLibrary;
-import io.ballerina.compiler.api.impl.types.util.BallerinaMethodDescriptor;
 import io.ballerina.compiler.api.symbols.Qualifier;
 import io.ballerina.compiler.api.types.BallerinaTypeDescriptor;
 import io.ballerina.compiler.api.types.FieldDescriptor;
 import io.ballerina.compiler.api.types.FunctionTypeDescriptor;
 import io.ballerina.compiler.api.types.Parameter;
-import io.ballerina.compiler.api.types.ParameterKind;
-import io.ballerina.compiler.api.types.util.MethodDescriptor;
+import io.ballerina.compiler.api.types.util.ParameterKind;
+import io.ballerina.compiler.api.types.MethodDescriptor;
 import org.ballerinalang.model.types.TypeKind;
 import org.wso2.ballerinalang.compiler.semantics.analyzer.Types;
 import org.wso2.ballerinalang.compiler.semantics.model.symbols.BAttachedFunction;
@@ -74,9 +73,9 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
-import static io.ballerina.compiler.api.types.ParameterKind.DEFAULTABLE;
-import static io.ballerina.compiler.api.types.ParameterKind.REQUIRED;
-import static io.ballerina.compiler.api.types.ParameterKind.REST;
+import static io.ballerina.compiler.api.types.util.ParameterKind.DEFAULTABLE;
+import static io.ballerina.compiler.api.types.util.ParameterKind.REQUIRED;
+import static io.ballerina.compiler.api.types.util.ParameterKind.REST;
 import static java.util.Collections.unmodifiableList;
 import static java.util.stream.Collectors.toList;
 import static org.ballerinalang.model.types.TypeKind.OBJECT;
