+++ tests/ballerina-compiler-api-test/src/test/java/io/ballerina/semantic/api/test/TypedescriptorTest.java
@@ -33,14 +33,14 @@ import io.ballerina.compiler.api.types.FutureTypeDescriptor;
 import io.ballerina.compiler.api.types.MapTypeDescriptor;
 import io.ballerina.compiler.api.types.ObjectTypeDescriptor;
 import io.ballerina.compiler.api.types.Parameter;
-import io.ballerina.compiler.api.types.ParameterKind;
+import io.ballerina.compiler.api.types.util.ParameterKind;
 import io.ballerina.compiler.api.types.RecordTypeDescriptor;
 import io.ballerina.compiler.api.types.TupleTypeDescriptor;
-import io.ballerina.compiler.api.types.TypeDescKind;
+import io.ballerina.compiler.api.types.util.TypeDescKind;
 import io.ballerina.compiler.api.types.TypeDescTypeDescriptor;
 import io.ballerina.compiler.api.types.TypeReferenceTypeDescriptor;
 import io.ballerina.compiler.api.types.UnionTypeDescriptor;
-import io.ballerina.compiler.api.types.util.MethodDescriptor;
+import io.ballerina.compiler.api.types.MethodDescriptor;
 import org.ballerinalang.test.util.CompileResult;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.DataProvider;
@@ -50,24 +50,24 @@ import org.wso2.ballerinalang.compiler.util.CompilerContext;
 
 import java.util.List;
 
-import static io.ballerina.compiler.api.types.ParameterKind.DEFAULTABLE;
-import static io.ballerina.compiler.api.types.ParameterKind.REQUIRED;
-import static io.ballerina.compiler.api.types.ParameterKind.REST;
-import static io.ballerina.compiler.api.types.TypeDescKind.ANYDATA;
-import static io.ballerina.compiler.api.types.TypeDescKind.ARRAY;
-import static io.ballerina.compiler.api.types.TypeDescKind.DECIMAL;
-import static io.ballerina.compiler.api.types.TypeDescKind.FLOAT;
-import static io.ballerina.compiler.api.types.TypeDescKind.FUTURE;
-import static io.ballerina.compiler.api.types.TypeDescKind.INT;
-import static io.ballerina.compiler.api.types.TypeDescKind.MAP;
-import static io.ballerina.compiler.api.types.TypeDescKind.NIL;
-import static io.ballerina.compiler.api.types.TypeDescKind.OBJECT;
-import static io.ballerina.compiler.api.types.TypeDescKind.RECORD;
-import static io.ballerina.compiler.api.types.TypeDescKind.STRING;
-import static io.ballerina.compiler.api.types.TypeDescKind.TUPLE;
-import static io.ballerina.compiler.api.types.TypeDescKind.TYPEDESC;
-import static io.ballerina.compiler.api.types.TypeDescKind.TYPE_REFERENCE;
-import static io.ballerina.compiler.api.types.TypeDescKind.UNION;
+import static io.ballerina.compiler.api.types.util.ParameterKind.DEFAULTABLE;
+import static io.ballerina.compiler.api.types.util.ParameterKind.REQUIRED;
+import static io.ballerina.compiler.api.types.util.ParameterKind.REST;
+import static io.ballerina.compiler.api.types.util.TypeDescKind.ANYDATA;
+import static io.ballerina.compiler.api.types.util.TypeDescKind.ARRAY;
+import static io.ballerina.compiler.api.types.util.TypeDescKind.DECIMAL;
+import static io.ballerina.compiler.api.types.util.TypeDescKind.FLOAT;
+import static io.ballerina.compiler.api.types.util.TypeDescKind.FUTURE;
+import static io.ballerina.compiler.api.types.util.TypeDescKind.INT;
+import static io.ballerina.compiler.api.types.util.TypeDescKind.MAP;
+import static io.ballerina.compiler.api.types.util.TypeDescKind.NIL;
+import static io.ballerina.compiler.api.types.util.TypeDescKind.OBJECT;
+import static io.ballerina.compiler.api.types.util.TypeDescKind.RECORD;
+import static io.ballerina.compiler.api.types.util.TypeDescKind.STRING;
+import static io.ballerina.compiler.api.types.util.TypeDescKind.TUPLE;
+import static io.ballerina.compiler.api.types.util.TypeDescKind.TYPEDESC;
+import static io.ballerina.compiler.api.types.util.TypeDescKind.TYPE_REFERENCE;
+import static io.ballerina.compiler.api.types.util.TypeDescKind.UNION;
 import static io.ballerina.semantic.api.test.util.SemanticAPITestUtils.compile;
 import static io.ballerina.tools.text.LinePosition.from;
 import static org.testng.Assert.assertEquals;
